refactor(session): split session module into separate files
Extract Session class:
- Move Session class implementation to session/_session.py
- Preserve singleton pattern and context manager functionality
- Keep check_call() and check_output() as Session methods

Extract module-level functions:
- Create session/_functions.py for open(), close(), check_call(), check_output()
- Implement convenience functions that work with singleton Session

Update module exports:
- Import and re-export from _session and _functions in __init__.py
- Update __all__ to include Session, functions, and error classes
- Remove inline implementation (reduced from 189 lines to imports)
