+++ vehicle-python/src/vehicle_lang/session/__init__.py
@@ -1,189 +1,16 @@
-import atexit
-import io
-import os
-import pty
-import sys
-from contextlib import AbstractContextManager, redirect_stderr, redirect_stdout
-from types import TracebackType
-from typing import TYPE_CHECKING, ClassVar, List, Optional, Sequence, Tuple, Type
+from typing import List
 
-from typing_extensions import Self, TypeAlias
-
-from .._binding import _unsafe_vehicle_free, _unsafe_vehicle_init, _unsafe_vehicle_main
-from .._temporary_files import temporary_files
-from ..error import VehicleError as VehicleError
-from ..typing import Target
+from ._functions import check_call, check_output, close, open
+from ._session import Session
 from .error import VehicleSessionClosed as VehicleSessionClosed
 from .error import VehicleSessionUsed as VehicleSessionUsed
 
 __all__: List[str] = [
-    "Target",
     "Session",
     "check_call",
     "check_output",
-    "loads",
-    "load",
+    "close",
+    "open",
+    "VehicleSessionClosed",
+    "VehicleSessionUsed",
 ]
-
-if TYPE_CHECKING or sys.version_info >= (3, 9):
-    SessionContextManager: TypeAlias = AbstractContextManager["Session"]
-else:
-    SessionContextManager: TypeAlias = AbstractContextManager
-
-
-class Session(SessionContextManager):
-    """
-    The Session class enforces that the Haskell RTS is only initialised once,
-    even when multiple calls to Vehicle are made.
-
-    This is needed as initialising the Haskell RTS multiple times is unsafe,
-    see: https://gitlab.haskell.org/ghc/ghc/-/issues/13693
-    """
-
-    _instance: ClassVar[Self]
-    _rts_init: bool
-    _rts_exit: bool
-
-    def __new__(cls: Type[Self]) -> "Session":
-        """
-        This override of __new__ enforces that Session is a singleton, i.e.,
-        that multiple calls to Session() return the same instance.
-        """
-        if not hasattr(cls, "_instance"):
-            cls._instance = super(Session, cls).__new__(cls)
-            cls._instance._rts_init = False
-            cls._instance._rts_exit = False
-        return cls._instance
-
-    def __enter__(self) -> Self:
-        if not self._rts_init:
-            self.open()
-        return self
-
-    def __exit__(
-        self,
-        _exc_type: Optional[Type[BaseException]],
-        _exc_value: Optional[BaseException],
-        _exc_traceback: Optional[TracebackType],
-    ) -> Optional[bool]:
-        if not self.closed:
-            self.close()
-        return None
-
-    def check_call(self, args: Sequence[str]) -> int:
-        if not self.closed:
-            return _unsafe_vehicle_main(args)
-        else:
-            raise VehicleSessionClosed()
-
-    def check_output(
-        self,
-        args: Sequence[str],
-    ) -> tuple[int, Optional[str], Optional[str], Optional[str]]:
-        with redirect_stdout(io.StringIO()) as out:
-            with redirect_stderr(io.StringIO()) as err:
-                with temporary_files("log", prefix="vehicle") as (log,):
-                    exitCode = self.check_call(
-                        [
-                            f"--redirect-logs={log}",
-                            *args,
-                        ]
-                    )
-                    return (
-                        exitCode,
-                        out.getvalue() or None,
-                        err.getvalue() or None,
-                        log.read_text(),
-                    )
-
-    def check_output_pty(
-        self, args: Sequence[str]
-    ) -> tuple[int, Optional[str], Optional[str], Optional[str]]:
-        stdout_fd = sys.stdout.fileno()
-        stderr_fd = sys.stderr.fileno()
-
-        # Create a pseudo-terminal pair to capture stdout
-        provider_fd, receiver_fd = pty.openpty()
-        # Create a pipe to capture stderr
-        pread_fd, pwrite_fd = os.pipe()
-
-        saved_stdout_fd = os.dup(stdout_fd)
-        saved_stderr_fd = os.dup(stderr_fd)
-
-        try:
-            # Redirect stdout and stderr
-            os.dup2(receiver_fd, stdout_fd)
-            os.dup2(pwrite_fd, stderr_fd)
-
-            # Close unused write ends
-            os.close(receiver_fd)
-            os.close(pwrite_fd)
-
-            with temporary_files("log", prefix="vehicle") as (log,):
-                exitCode = self.check_call([f"--redirect-logs={log}", *args])
-
-            sys.stdout.flush()
-
-        finally:
-            # Restore stdout and stderr
-            os.dup2(saved_stdout_fd, stdout_fd)
-            os.dup2(saved_stderr_fd, stderr_fd)
-
-            # Cleanup
-            os.close(saved_stdout_fd)
-            os.close(saved_stderr_fd)
-
-        out_lines = []
-
-        with os.fdopen(provider_fd, "r") as f:
-            try:
-                while line := f.readline():
-                    out_lines.append(line)
-            except OSError:
-                pass
-
-        with os.fdopen(pread_fd, "r") as f:
-            err = f.read()
-
-        return (
-            exitCode,
-            "".join(out_lines) or None,
-            err or None,
-            log.read_text(),
-        )
-
-    def close(self) -> None:
-        if not self.closed:
-            self._rts_exit = True
-            _unsafe_vehicle_free()
-            atexit.unregister(self.close)
-
-    @property
-    def closed(self) -> bool:
-        return not self._rts_init or self._rts_exit
-
-    def open(self, rts_args: Optional[Sequence[str]] = None) -> None:
-        if self._rts_init:
-            raise VehicleSessionUsed()
-        else:
-            self._rts_init = True
-            _unsafe_vehicle_init(["vehicle", *(rts_args or [])])
-            atexit.register(self.close)
-
-
-def check_call(args: Sequence[str]) -> int:
-    return Session().__enter__().check_call(args)
-
-
-def check_output(
-    args: Sequence[str],
-) -> tuple[int, Optional[str], Optional[str], Optional[str]]:
-    return Session().__enter__().check_output_pty(args)
-
-
-def close() -> None:
-    Session().close()
-
-
-def open(rts_args: Optional[Sequence[str]] = None) -> None:
-    Session().open(rts_args)
